+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Compatibility:
 * Handle `Kernel#clone(freeze: true)` (#2512, @andrykonchin).
 * Relax `Fiber#transfer` limitations (#2453).
 * Implement `Fiber#blocking?` like CRuby 3 (#2453, @aardvark179).
+* Implement `rb_str_locktmp` and `rb_str_unlocktmp` (#2524, @bjfish).
 
 Performance: